changes in GDD, starting new game bug and salesman teleporting
Added already reached teleport points as data to be saved to GDD and implemented it.

A new game can be started from a running one without crashing now. But there's a new problem with marking the interaction with NPCs. Needs fixing.

Salesman teleport is now possible between already reached level, there's no need to talk to all the salesmen on the way anymore.